support for Tailer config mixin
# Conflicts:
#	bin/src/main.rs
#	common/fs/src/cache/dir_path.rs
