+++ bin/src/main.rs
@@ -1,8 +1,8 @@
 use config::{self, Config};
 use std::sync::Arc;
 use tokio::sync::Mutex;
-
-#[cfg(unix)]
+use tracing::info;
+#[cfg(not(windows))]
 use tracing::{debug, info, trace, warn};
 
 #[cfg(windows)]
